services/report-generator/t3_report_generator/pyproject.toml
[project]
name = "t3-report-generator"
version = "0.2.0"
description = "AI-driven PDF report generator for CAPA ad logs"
requires-python = ">=3.12"
dependencies = [
    "slack-bolt>=1.18.0",
    "anthropic>=0.40.0",
    "pandas>=2.0.0",
    "matplotlib>=3.8.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "reportlab>=4.0.0",
    "markdown>=3.7",
]
